simulation/requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
scipy>=1.11.0
matplotlib>=3.7.0
numpy>=1.24.0
